api/serializers/cliente_serializer.py
from rest_framework import serializers
from api.models.cliente import Cliente

class ClienteSerializer(serializers.ModelSerializer):
    class Meta:
        model = Cliente
        fields = ['id', 'nombre', 'canal'] 